Fix compiler warning: implicit conversion loses integer precision
grisbi/src/utils_str.c:728:15: warning: implicit conversion loses integer precision: 'unsigned long' to 'gint' (aka 'int') [-Wshorten-64-to-32]
    str_len = strlen ( str );
            ~ ^~~~~~~~~~~~~~
grisbi/src/utils_str.c:729:16: warning: implicit conversion loses integer precision: 'unsigned long' to 'gint' (aka 'int') [-Wshorten-64-to-32]
    long_old = strlen ( old_str );
             ~ ^~~~~~~~~~~~~~~~~~

+++ src/utils_str.c
@@ -718,7 +718,7 @@ gchar *gsb_string_remplace_string ( const gchar *str,
                         gchar *new_str )
 {
     gchar *ptr_debut;
-    gint long_old, str_len;
+    size_t long_old, str_len;
     gchar *chaine, *ret, *tail;
 
     ptr_debut = g_strstr_len ( str, -1, old_str);
